backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
anthropic>=0.34.0
openai>=1.0.0
requests>=2.31.0
weasyprint>=61.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
pytest>=7.4.0
pytest-cov>=4.1.0
